+++ src/env/node/git/localGitProvider.ts
@@ -207,7 +207,7 @@ interface RepositoryInfo {
 
 export class LocalGitProvider implements GitProvider, Disposable {
 	readonly descriptor: GitProviderDescriptor = { id: GitProviderId.Git, name: 'Git', virtual: false };
-	readonly supportedSchemes: Set<string> = new Set([
+	readonly supportedSchemes = new Set<string>([
 		Schemes.File,
 		Schemes.Git,
 		Schemes.GitLens,
@@ -571,12 +571,14 @@ export class LocalGitProvider implements GitProvider, Disposable {
 			: [RepositoryVisibility.Private, getVisibilityCacheKey(remotes)];
 	}
 
+	private _pendingRemoteVisibility = new Map<string, ReturnType<typeof fetch>>();
 	@debug<LocalGitProvider['getRemoteVisibility']>({ args: { 0: r => r.url }, exit: r => `returned ${r[0]}` })
 	private async getRemoteVisibility(
 		remote: GitRemote,
 	): Promise<[visibility: RepositoryVisibility, remote: GitRemote]> {
 		const scope = getLogScope();
 
+		let url;
 		switch (remote.provider?.id) {
 			case 'github':
 			case 'gitlab':
@@ -584,27 +586,45 @@ export class LocalGitProvider implements GitProvider, Disposable {
 			case 'azure-devops':
 			case 'gitea':
 			case 'gerrit':
-			case 'google-source': {
-				const url = remote.provider.url({ type: RemoteResourceType.Repo });
+			case 'google-source':
+				url = remote.provider.url({ type: RemoteResourceType.Repo });
 				if (url == null) return [RepositoryVisibility.Private, remote];
 
-				// Check if the url returns a 200 status code
-				try {
-					const rsp = await fetch(url, { method: 'HEAD', agent: getProxyAgent() });
-					if (rsp.ok) return [RepositoryVisibility.Public, remote];
-
-					Logger.debug(scope, `Response=${rsp.status}`);
-				} catch (ex) {
-					debugger;
-					Logger.error(ex, scope);
+				break;
+			default: {
+				url = remote.url;
+				if (!url.includes('git@')) {
+					return maybeUri(url)
+						? [RepositoryVisibility.Private, remote]
+						: [RepositoryVisibility.Local, remote];
 				}
-				return [RepositoryVisibility.Private, remote];
+
+				const [host, repo] = url.split('@')[1].split(':');
+				if (!host || !repo) return [RepositoryVisibility.Private, remote];
+
+				url = `https://${host}/${repo}`;
 			}
-			default:
-				return maybeUri(remote.url)
-					? [RepositoryVisibility.Private, remote]
-					: [RepositoryVisibility.Local, remote];
 		}
+
+		// Check if the url returns a 200 status code
+		let promise = this._pendingRemoteVisibility.get(url);
+		if (promise == null) {
+			promise = fetch(url, { method: 'HEAD', agent: getProxyAgent() });
+			this._pendingRemoteVisibility.set(url, promise);
+		}
+
+		try {
+			const rsp = await promise;
+			if (rsp.ok) return [RepositoryVisibility.Public, remote];
+
+			Logger.debug(scope, `Response=${rsp.status}`);
+		} catch (ex) {
+			debugger;
+			Logger.error(ex, scope);
+		} finally {
+			this._pendingRemoteVisibility.delete(url);
+		}
+		return [RepositoryVisibility.Private, remote];
 	}
 
 	@log<LocalGitProvider['repositorySearch']>({
@@ -4094,10 +4114,9 @@ export class LocalGitProvider implements GitProvider, Disposable {
 		const data = await this.git.status__file(root, relativePath, porcelainVersion, {
 			similarityThreshold: configuration.get('advanced.similarityThreshold'),
 		});
-		const status = GitStatusParser.parse(data, root, porcelainVersion);
-		if (status == null || !status.files.length) return undefined;
 
-		return status.files[0];
+		const status = GitStatusParser.parse(data, root, porcelainVersion);
+		return status?.files?.[0];
 	}
 
 	@log()
@@ -4109,10 +4128,9 @@ export class LocalGitProvider implements GitProvider, Disposable {
 		const data = await this.git.status__file(root, relativePath, porcelainVersion, {
 			similarityThreshold: configuration.get('advanced.similarityThreshold'),
 		});
-		const status = GitStatusParser.parse(data, root, porcelainVersion);
-		if (status == null || !status.files.length) return [];
 
-		return status.files;
+		const status = GitStatusParser.parse(data, root, porcelainVersion);
+		return status?.files ?? [];
 	}
 
 	@log()