+++ src/plus/workspaces/models.ts
@@ -7,15 +7,15 @@ export enum WorkspaceType {
 	Cloud = 'cloud',
 }
 
-export enum WorkspaceSyncSetting {
-	Never = 'never',
-	Always = 'always',
-	Ask = 'ask',
+export enum WorkspaceAutoAddSetting {
+	Disabled = 'disabled',
+	Enabled = 'enabled',
+	Prompt = 'prompt',
 }
 
 export type CodeWorkspaceFileContents = {
 	folders: { path: string }[];
-	settings: { [key: string]: any };
+	settings: Record<string, any>;
 };
 
 export type WorkspaceRepositoriesByName = Map<string, RepositoryMatch>;
@@ -29,6 +29,7 @@ export interface RemoteDescriptor {
 	provider: string;
 	owner: string;
 	repoName: string;
+	url?: string;
 }
 
 export interface GetWorkspacesResponse {
@@ -414,6 +415,12 @@ interface CloudWorkspaceFetchedConnection<i> extends CloudWorkspaceConnection<i>
 	is_fetching: boolean;
 }
 
+export interface WorkspaceResponse {
+	data: {
+		project: CloudWorkspaceData;
+	};
+}
+
 export interface WorkspacesResponse {
 	data: {
 		projects: CloudWorkspaceConnection<CloudWorkspaceData>;
@@ -484,9 +491,7 @@ export type AddRepositoriesToWorkspaceResponse = {
 	data: {
 		add_repositories_to_project: {
 			id: string;
-			provider_data: {
-				[repoKey: string]: CloudWorkspaceRepositoryData;
-			};
+			provider_data: Record<string, CloudWorkspaceRepositoryData>;
 		} | null;
 	};
 	errors?: { code: number; message: string }[];
@@ -576,9 +581,7 @@ export interface LocalWorkspaceFileData {
 	workspaces: LocalWorkspaceData;
 }
 
-export type LocalWorkspaceData = {
-	[localWorkspaceId: string]: LocalWorkspaceDescriptor;
-};
+export type LocalWorkspaceData = Record<string, LocalWorkspaceDescriptor>;
 
 export interface LocalWorkspaceDescriptor {
 	localId: string;
@@ -603,19 +606,13 @@ export interface CloudWorkspaceFileData {
 	workspaces: CloudWorkspacesPathMap;
 }
 
-export type CloudWorkspacesPathMap = {
-	[cloudWorkspaceId: string]: CloudWorkspacePaths;
-};
+export type CloudWorkspacesPathMap = Record<string, CloudWorkspacePaths>;
 
 export interface CloudWorkspacePaths {
 	repoPaths: CloudWorkspaceRepoPathMap;
 	externalLinks: CloudWorkspaceExternalLinkMap;
 }
 
-export type CloudWorkspaceRepoPathMap = {
-	[repoId: string]: string;
-};
+export type CloudWorkspaceRepoPathMap = Record<string, string>;
 
-export type CloudWorkspaceExternalLinkMap = {
-	[fileExtenstion: string]: string;
-};
+export type CloudWorkspaceExternalLinkMap = Record<string, string>;